Adjusted dupe and fingerprint handling in WebFormAction
Rearranged duplicate and fingerprint handling in WebFormAction's contact
form submission. Previously, a user submitting the contact form
who is already associated with a Fingerprint for a Contact but provides
different details, such as email, would result in a new Contact being
created, but with the same fingerprintId. Hence, the constraint in the
database would prevent the Contact from being saved. Furthermore, the
form UI behaves inappropriately in this case, reloading without error
messages.

Now the system will be checked for duplicates, then subsequently looked
up by fingerprint (not partial match) to prevent this behavior. As
before, the input provided in the form is merged in with the existing
Contact.