Added dropdown to add Actions to a calendar
Since the calendar overhaul in 6.0, the calendar has been restricted
to Actions with no type or type 'event'. However, since the calendarId
field would not be set when creating actions, no standard actions would be
present on the calendar. A dropdown has been added to the actions/create
page and the Action tab of the publisher to select the associated calendar.