Improvements to header format handling in EmailDeliveryBehavior::addressHeaderToArray
To improve compatibility with email address header formats, additional
checks have been added to addressHeaderToArray when initial validation
fails. This will use the next piece of the $headerArray to see if we've
inadvertantly split in an incorrect position, due to use of single quotes,
or embedded single quotes, e.g. "'Test Contact'".