Added reverse geocoding request to front end of checkins
When a Google API key is given a referrer restriction, it can no longer
be used server-side for the web service APIs like geocoding. The
geocoding first takes place client side (when using geolocation), and
will be performed server side as fallback. Failure to issue the API call
is now properly logged instead of causing a failing ajax request. Should
all attempts to retrieve an address fail, that part of the message is
left off