ActionOverdueTrigger condition bugfix
Updated ActionOverdueTrigger and X2FlowItem to allow and use a
'defaultOperator' parameter. Previously, the default operator = would
have been used for the comparison, as the operator selector would not be
present in the UI, and therefore not be included in the options. In the
event that cron was NOT configured to run every minute, this =
comparison would most likely fail.

Furthermore, the static function checkCondition in ActionOverdueTrigger
would never fire, as this method is not called by X2FlowTrigger for the
specified option type. This method has been removed.